[ios] Replace MultiProfileSignOut(...) API by ProfileSignoutRequest{}
The clients of MultiProfileSignOut(...) may have to behave differently
if the sign-out operation requires changing the Profile. To propagate
this information a new optional parameter would have to be passed (a
callback invoked with a boolean indicating whether the profile would
be changed or not).

This caused the function to have most of its parameters optional and
invocation were difficult to parse (due to the noise of passing all
the optional parameter).

Instead change the API to use a class ProfileSignoutRequest{} that
follow the builder pattern to set the optional parameter. Once it is
initialised, the sign-out operation can be started by calling the
Run(...) method with the Browser.

Remove some unnecessary conversion between callback and blocks.

Bug: 407697351
Change-Id: If92282b32bfda8645496d8a07746a7cb0fe19609
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6416885
Reviewed-by: Marc Treib <treib@chromium.org>
Commit-Queue: Sylvain Defresne <sdefresne@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1443391}