[InputVizard] Add crash keys to log operations on frame sink.
Using this crash key we would like to get context on crash happening,
mainly trying to see if a RegisterFrameSinkHierarchy call comes before
CreateCompositorFrameSink. In the current implementation in
InputManager, it is assumed RegisterFrameSinkHierarchy always comes
after CreateCompositorFrameSink. In case of this out of ordered call
RendererInputRouterSupports are created with incorrect types.


Bug: 404741207
Change-Id: Iabd622f5878a830b0c682666762dc5fa71b752a7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6430109
Reviewed-by: Aman Verma <amanvr@google.com>
Commit-Queue: Kartar Singh <kartarsingh@google.com>
Cr-Commit-Position: refs/heads/main@{#1443393}