[//cc] Remove TileDrawInfo::is_premultiplied_
This field is effectively no longer used. It is initialized to false and
is also set to false here [1] and is set to true in `SetResource()`.
However, in reality all callsites of
`TileDrawInfo::is_premultiplied` require that the resource be present as
they call `resource_id_for_export()` [2]. TileDrawInfo has a resource
only after a call to `SetResource()` has been made without a following
call to `TakeResource()`.

[1] https://source.chromium.org/chromium/chromium/src/+/main:cc/tiles/tile_draw_info.cc;l=49;drc=78883e6c824fd669688b9943096a7b01f87c1cb5;bpv=1;bpt=1
[2]
https://source.chromium.org/chromium/chromium/src/+/main:cc/tiles/tile_draw_info.h;l=51;drc=78883e6c824fd669688b9943096a7b01f87c1cb5;bpv=1;bpt=1

Bug: 403372453
Change-Id: I8c0deb294f37fee8f2d87327c3ee783021e17ea8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6432463
Reviewed-by: Vasiliy Telezhnikov <vasilyt@chromium.org>
Commit-Queue: Colin Blundell <blundell@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1443416}