[ios] Unload profiles when last scene is disconnected
Expose a new observer method -profileState:sceneDisconnected: to
ProfileStateObserver and use it in MainController to schedule an
unload for the profile when they no longer have any connected
scenes.

Using this new API, the profile will automatically be unloaded
after a profile is deleted via -deleteProfile: as all the scenes
will be re-connected to the personal profile, thus the logic in
-deleteProfile: can be simplified to only schedule the profile
switch.

The profile is not immediately disconnected when the last scene
is disconnected (to avoid destroying the object notifying its
observers) but instead the unloading is scheduled after to happen
in the next execution of the main loop.

Fixed: 408198647
Change-Id: I25e08a3950f63cffa96dd0e893b8e44de7dd503f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6433933
Commit-Queue: Sylvain Defresne <sdefresne@chromium.org>
Reviewed-by: Marc Treib <treib@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1443424}