Add 5 minute reporting of Skia's Graphite Pipeline caching
This will surface the Skia UMA stat:
   Skia.Graphite.PipelineCache.PipelineUsesInEpoch
which reports the number of Pipelines used since the last call to reportPipelineStats(kPipelineCache). We will use this to set Skia's constant:
   kGlobalGraphicsPipelineCacheSizeLimit
which is currently set (somewhat arbitrarily) to 256.

Low-Coverage-Reason: EXPERIMENTAL_CODE Precompilation is not yet enabled
Bug: 358074434
Change-Id: Ic2a37bb0062bfa54553ae232492604f995e2e5d5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6426285
Commit-Queue: Robert Phillips <robertphillips@google.com>
Reviewed-by: Vasiliy Telezhnikov <vasilyt@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1443500}