Remove usage of SKPath in CanvasPathTest
In this refactor, SKPath is used to store paths and PathBuilder is used
to construct paths; and the path construction will be removed from
SkPath. In the previous cl, CanvasPath is updated to use PathBuilder.
Update the CanvasPathTest to compare its results with PathBuilder
results to reflect the refactor.

Bug: 378688986

Change-Id: Ieaec9ff2b68b78b5dfe09491cbfa24186749f002
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6433404
Reviewed-by: Florin Malita <fmalita@chromium.org>
Commit-Queue: Yi Xu <yiyix@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1443502}