Serialize SkPath with mojo
This is required for making corner-shape and backdrop-filter work well
together (the path need to be propagated all the way to viz).

This uses the internal serialization mechanism of SkPath, and saves it
into a mojo array<uint8>. The path is "opaque" but can be serialized
to/from an SkPath.

Bug: 408498304
Change-Id: I606dc0e78789cda1552bcd4c55a84958b5ba3a91
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6434053
Reviewed-by: Robert Phillips <robertphillips@google.com>
Reviewed-by: Mike West <mkwst@chromium.org>
Commit-Queue: Noam Rosenthal <nrosenthal@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1443511}