Add Desk sync section to the chrome://management page
A new section for the Desk sync feature which will be shown on ChromeOS
as long as either windows and/or cookies are synced for the managed
user. Implemented as in the UI mock which is linked in the bug.

Tests in management_ui_handler_unittest.cc had an unclear dependency
between users and profiles and never reported test user as active, which
blocked detection of Floating Workspace. To resolve this, they were
rewritten to use TestingProfileManager.

Bug: b:376038342
Change-Id: I325bbd675a1f368b7df3ce6fb892ef36edade027
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6424430
Commit-Queue: Andrey Davydov <andreydav@google.com>
Reviewed-by: Artem Sumaneev <asumaneev@google.com>
Reviewed-by: Cait Phillips <caitkp@google.com>
Cr-Commit-Position: refs/heads/main@{#1443513}