[iOS] NTP: Fix "Sign in" briefly flashing while switching accounts
When switching accounts (with chrome://flags/#identity-disc-account-menu
and/or chrome://flags/#separate-profiles-for-managed-accounts enabled),
there was a brief flash of the "Sign in" button between the old and the
new avatar image.
This was previously fixed by listening to
onEndBatchOfPrimaryAccountChanges (which only gets called once both the
sign-out and the sign-in are complete), see bug 397194715, but
https://crrev.com/c/6355423 reintroduced the flashes by adding an extra
setUserSignedIn notification which bypassed the IdentityManager-level
batching.
This CL fixes the issue by removing setUserSignedIn, which wasn't
really necessary in the first place - the signin state could already
be determined from existing calls.

Fixed: 406922383
Change-Id: Id602f0103fbc1019a9b84735441daec1796870bf
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6433737
Commit-Queue: Marc Treib <treib@chromium.org>
Reviewed-by: Scott Yoder <scottyoder@google.com>
Cr-Commit-Position: refs/heads/main@{#1443518}