[Blob URL] Extend cross-partition blob URL fetching bypass to dedicated workers
If a document context requests storage access, has it granted, and then
creates a dedicated worker, that dedicated worker should be considered
to have storage access as well. This means that if the third-party
context creates a blob URL using the StorageAccessHandle and sends it to
the dedicated worker, the dedicated worker should be able to fetch it.

Bug: 403297818
Change-Id: Ifebc1f5ec78bd0e8e79e48ed21f6c98b90def715
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6409631
Reviewed-by: Bo Liu <boliu@chromium.org>
Reviewed-by: Fergal Daly <fergal@chromium.org>
Reviewed-by: Andrew Williams <awillia@chromium.org>
Reviewed-by: Robert Flack <flackr@chromium.org>
Commit-Queue: Janice Liu <janiceliu@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1443531}