Make category param name configurable for BeaconLeakageLogging
Currently the category param name used is hardcoded to "category", but
there is a request to allow this to be configurable. This CL makes it
so that the param name itself is configurable via Finch, but still
defaults to "category" if unset.

Bug: 382527001
Change-Id: If845cdc9ae8b17fab0026dde14578479edfce702
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6432811
Commit-Queue: Rakina Zata Amni <rakina@chromium.org>
Reviewed-by: Ian Clelland <iclelland@chromium.org>
Commit-Queue: Ian Clelland <iclelland@chromium.org>
Auto-Submit: Rakina Zata Amni <rakina@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1443597}