[EnterprisePill] Add profile management check
The profile menu crashes when it is shown and the user has previously
accepted management but the profile is not managed. It happens when
signing in to a managed account which has management disabled. This
updates the behavior to only show the management UI when the profile is
managed and not only the account.

The CL also removes the default enterprise profile name "Work" even when
the profile is managed, but keeps the profile name shown as the custom
label when the `EnterpriseCustomLabel` policy is set.

A follow-up CL will add an observer to listen to profile management
changes to update the avatar to/away from the default "Work" label.

Bug: 406511295
Change-Id: I57cda00a5a4a3adb114f661997b585be416f2206
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6428737
Reviewed-by: Yann Dago <ydago@chromium.org>
Reviewed-by: David Roger <droger@chromium.org>
Commit-Queue: Salma Elmahallawy <esalma@google.com>
Cr-Commit-Position: refs/heads/main@{#1443603}