Fix k-anonymity key caching when all keys are fetched from the cache.
In https://crrev.com/c/6361984, we added caching of keys fetched from
the k-anonymity server, guarded by the `FledgeCacheKAnonHashedKeys`
feature. This has not yet been enabled on any channel.

We since discovered that this change leaks `InProgressQueryState` in the
`InterestGroupKAnonymityManager`'s queries_in_progress_ map whenever all
of the keys being fetched are available from the cache, and also, the
interest group's record of k-anonymous keys is not updated to reflect
that query. This change fixes both of this issues by specifically
handling the case where all of the keys are retrieved from the cache.

Change-Id: I05966330203069c2463b109b3c37685d5b43aab1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6434042
Commit-Queue: Orr Bernstein <orrb@google.com>
Reviewed-by: Russ Hamilton <behamilton@google.com>
Cr-Commit-Position: refs/heads/main@{#1443646}