Remove histograms_allowlist.txt and access histogram names in Java
Though Resource$NotFoundExceptions have decreased by a good amount, we
still see a lot of crashes originating from attempting to read the
histograms allowlist. Based on our data, we believe this accounts for
almost 1 million daily crashes.

This CL gets rid of the histograms_allowlist text file which defines the
histograms which are allowed to sample at 100% and instead moves the
histograms names into the HistogramAllowlist class.

This CL also updates the presubmit checks which ensure that the
histogram names exist in our code. Where before, the Python check would
read the text file, it now parses the list directly from the Java file
by looking for the START_PARSING and END_PARSING markers either side of
the content of the list.

Bug: 332746920
Change-Id: I9ace762e6c18223c4d2d05921bc743d80dbc8cad
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6432446
Reviewed-by: Nate Fischer <ntfschr@chromium.org>
Commit-Queue: Alex Mitra <alexmitra@chromium.org>
Reviewed-by: Robert Kaplow <rkaplow@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1443705}