Assert that WebUI JS test name length doesn't exceed ResultDB limit
The goal is to have WebUI JS tests fail locally if the test name is long
enough to exceed ResultDB's 512 byte limit on test identifiers.

The test identifier consumed by result_adapter is constructed in
//base/test/launcher/test_results_tracker.cc, which produces the JSON
that is consumed by result_adapter. This poses two problems:

  1. Failure at this stage is undetectable outside of a LUCI context
     (i.e. in local test runs)
  2. GTest code can't know how the final test identifier is constructed

To achieve the stated goal in light of problem 1, we must add an
assertion in WebUIMochaBrowserTest so that the test will fail locally if
the name is going to exceed the 512 byte limit imposed by ResultDB.

Because of problem 2, this guardrail can't be exact, and may become
stale.

These shortcomings are acknowledged in order to attain fast failure.

Bug: 393139276
Change-Id: Id4fde82a144c40a7f80e4071b710188bc85b880a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6384039
Reviewed-by: Demetrios Papadopoulos <dpapad@chromium.org>
Commit-Queue: Alex Yang <aycyang@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1443725}