[omnibox] Fix most visited deduping with open tabs.
This makes a change to the TabMatcher interface that allows it to
check for other tabs with the same title or the same url after a
set of replacements. We choose to only replace the ref because two
equivalent URLs after ref replacement is a strong indication they're the
same page.

The title itself is not a strong enough indication since events like
page loading can cause the title of the page to change. In events like
this, checking if the URLs are the same without the ref is a
strong fallback option. Also, even if pages don't have the same title
if the URLs  are the same after ref replacement, it is once again a strong indication that they are still the same page.
page.

Change-Id: I0c268fc4258b530973b62145cbc2f03e5a9d5ae8
Bug: 407554265, 406503169
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6426085
Reviewed-by: Moe Ahmadi <mahmadi@chromium.org>
Commit-Queue: Nihar Majmudar <niharm@google.com>
Cr-Commit-Position: refs/heads/main@{#1443760}