Provide dbus appmenu information on Wayland
The appmenu protocol provides a way to associate a wl_surface with a
global application menu, which uses the com.canonical.dbusmenu
interface. It is a KDE protocol, but it is also used by Qt (and Firefox
started using it recently too, although the corresponding feature is
hidden behind a flag).

Change-Id: I8d0a52b35ffd4f9c0d98469229fff8e019fca151
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6405535
Reviewed-by: Thomas Anderson <thomasanderson@chromium.org>
Reviewed-by: Nick Yamane <nickdiego@igalia.com>
Commit-Queue: Thomas Anderson <thomasanderson@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1443778}