[OmniboxAPI] Adding custom icon support in omnibox actions
This CL adds the ability to set a custom icon in suggestion actions. The
icon is set the same way as action icons, and reuses parsing methods.
This logic will also be reused when the suggestion.icon_url is migrated
to ImageData as well.

Recording:
https://drive.google.com/file/d/1npcj6ikHjctJb8ydNpPASdtLDTcXZA8W/view?usp=sharing&t=1

Bug: 381279444
Change-Id: If1e504f0a4cd4e7811b2ece341844dca84f0fdef
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6384375
Reviewed-by: Devlin Cronin <rdevlin.cronin@chromium.org>
Reviewed-by: Moe Ahmadi <mahmadi@chromium.org>
Commit-Queue: Salma Elmahallawy <esalma@google.com>
Cr-Commit-Position: refs/heads/main@{#1443799}