MemorySaver: Return Interactive tests to using NTP in most cases
In these tab discard test, a seemingly arbitrary second tab is
spawned, such that the original tab can be discarded. The URL of the
second tab was changed from NTP to Version (which loads faster), but
this appears to have exposeed a race condition.

Restore all but one test to using NTP, as they did before.  The one
test that remains using the Version page actually tests that it can
discard that page, so it can be treated separately.  This lets us
enable the disabled flaky test, buying time to fix this properly.

Bug: 408399396

Change-Id: I4a2c8996a3fc7aa0b32897aa7fb92bdacbb9e95e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6438500
Reviewed-by: David Pennington <dpenning@chromium.org>
Commit-Queue: Christopher Grant <cjgrant@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1443837}