[ServiceWorker] Add a sync WillConstructURLLoaderFactory observer method
This method allows extensions to get a notice that the URLLoaderFactory
that will be used to fetch the extension's worker script is about to be
constructed.

This is useful to prevent WebRequestAPI extensions from calling
ResetURLLoaderFactories while worker script fetching is in progress,
which can cause service worker registrations to fail.

In addition, this method is necessary to deterministically trigger
the race condition in browsertests.

Design doc: http://go/defer-reset-url-loader-factories

Bug: 394523691
Change-Id: Iadfbcd1b76dbb7906a911f538f51b3e73b4de691
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6395758
Reviewed-by: Yoshisato Yanagisawa <yyanagisawa@chromium.org>
Commit-Queue: Andrea Orru <andreaorru@chromium.org>
Reviewed-by: Shunya Shishido <sisidovski@chromium.org>
Reviewed-by: Kinuko Yasuda <kinuko@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1443858}