Reland of "Add UMA for the delay of ... and connected delegate"
This is a reland of https://crrev.com/c/6415568, focusing only UMA for a
connected callback to make a reland CL smaller and isolated.

The original CL causes a crash (https://crbug.com/407971829). The
underlying cause was that `delegate_connected_end_time_` was not
cleared, even though `delegate_->OnConnected` could be called multiple
times on the same instance (see also: https://crrev.com/c/6424639).

Given the complexity of the state machine, relying on its understanding
to determine when to reset the times seems risky.

Therefore, this CL now manages `connected_callback_{start,end}_time`
within the HTTP network transaction, instead of the URL request. The
HTTP network transaction is the lowest layer that invokes the connected
callback.

This CL also makes sure to reset `end_time` when setting `start_time` to
prevent an inconsistent state, which may cause a crash unexpectedly.

Note this is a kind of speculative fix because I couldn't reproduce a
crash locally.

Bug: 393980912, 407971829

Change-Id: Ie416ef14f6190e68fdf6fe32630fdd8a1aa3e47d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6434248
Reviewed-by: Kenichi Ishibashi <bashi@chromium.org>
Commit-Queue: Hayato Ito <hayato@chromium.org>
Reviewed-by: Dave Tapuska <dtapuska@chromium.org>
Reviewed-by: Takashi Toyoshima <toyoshim@chromium.org>
Reviewed-by: Kouhei Ueno <kouhei@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1443892}