[DIPS] Don't RazeAndPoison sqlite db if open failed
Failing to open the sqlite db can result in an error callback while
the internal db handle is null.
Calling RazeAndPoision if the internal db handle is null will cause a
failed DCHECK. To avoid that bail out of the error handler if the
database is not open.

Bug: 402564638

Change-Id: Ic350ff3b99b754d88a9902777c2bca1696a9a624
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6333358
Reviewed-by: Joshua Hood <jdh@chromium.org>
Reviewed-by: Bo Liu <boliu@chromium.org>
Commit-Queue: David Landell <david.landell@xperi.com>
Cr-Commit-Position: refs/heads/main@{#1443960}