[//cc] Remove TileDrawQuad taking in `is_premultiplied`
All production calls explicitly pass `true` except for this one [1]. In
that one, the value will have come from a user-passed dict. However, in
reality it is *not* a production call but rather just looks like one:
All callsites either are explicitly in tests or bottom out in this
method [2], which is itself only called in tests.

This CL removes passing the field from all test contexts. Some of those
test contexts currently pass false, but (a) there's no reason for them
to do that when production always passes `true`, and (b) the bots are
green.

[1] https://source.chromium.org/chromium/chromium/src/+/main:components/viz/common/quads/render_pass_io.cc;l=1493-1498;drc=78883e6c824fd669688b9943096a7b01f87c1cb5;bpv=1;bpt=1
[2] https://source.chromium.org/chromium/chromium/src/+/main:components/viz/common/quads/render_pass_io.cc;l=2123;drc=78883e6c824fd669688b9943096a7b01f87c1cb5

Bug: 403372453
Change-Id: I0896dd4f87d03915e6c8a0e65e739be5994775fb
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6434490
Reviewed-by: Vasiliy Telezhnikov <vasilyt@chromium.org>
Commit-Queue: Colin Blundell <blundell@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1443966}