Add WebExposed and AwParameterizedTest coverage for WebView Payments APIs
The PaymentRequest API is disabled by default in WebView until an app
(or test) turns it on via setPaymentRequestEnabled. This change turns
it on for WebExposedTest and AwParameterizedTest to ensure coverage.

Unstable WebExposed expectations are updated with the added coverage.

The feature is not yet stable for WebView, so stable expectations will
remain unchanged until the fieldtrial_testing_config.json entry is
removed.

Bug: 409053958
Change-Id: I6400e7a5a1298ba74fc0980581aa4b35a65dec3a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6439011
Auto-Submit: Ashley Newson <ashleynewson@chromium.org>
Reviewed-by: Rouslan Solomakhin <rouslan@chromium.org>
Reviewed-by: Mateusz Zbikowski <zbikowski@google.com>
Reviewed-by: Nate Fischer <ntfschr@chromium.org>
Commit-Queue: Ashley Newson <ashleynewson@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1444021}