[LHS Indicator] Update metrics and skip data with interval no more than 4s
This CL updates indicators tracking logic to avoid recording how
often an indicator is shown for no more than 4 seconds. The indicator
minimal show duration is 4 seconds, in other words it will never be
shown less than or equal to 4 seconds and we do not need to record it.

Bug: 407968839
Change-Id: I4069d3ccc1b08adbd237723c94f3ad1d2c776ee9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6419251
Commit-Queue: Yifan Luo <lyf@chromium.org>
Reviewed-by: Ravjit Uppal <ravjit@chromium.org>
Reviewed-by: Elias Klim <elklm@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1444077}