[AppIntegration] Fix crash due to re-register observer.
The AuxiliarySearchProvider is a KeyedService (singleton) and
registers itself when setObserver() is called. The crash was due to
multiple instances of AuxiliarySearchBridge (in Java) call AuxiliarySearchProvider#SetObserver() when it has been registered. Add
an parameter |is_observing_| and don't register again prevents the
crash.

Bug: 409115968
Change-Id: I7639b9c64c694d43c367b3a7ba9dcb7a51c34fe1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6437615
Commit-Queue: Xi Han <hanxi@chromium.org>
Reviewed-by: Gang Wu <gangwu@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1444089}