Make WebResourceResponse use JniZero
This CL updates WebResourceResponseInfo with JniType annotations to make
unpacking the object on the native side cleaner.

As part of this, the check to ensure that the Java InputStream is only
transferred to native once is moved to the Java side, where the java
reference is also reset to null, so subsequent callers will be unable to
access the stream.

This is not a perfect guarantee that the stream won't be used by other
callers on the Java side, but it does clarify the semantics of
transferring the stream to native code.

Bug: 364913812
Bug: 409249046
Change-Id: I81062026bebe50aa9afc7bde57a0136ae5ac5795
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6424427
Auto-Submit: Peter Pakkenberg <pbirk@chromium.org>
Commit-Queue: Peter Pakkenberg <pbirk@chromium.org>
Reviewed-by: Colin Blundell <blundell@chromium.org>
Reviewed-by: Peter Conn <peconn@chromium.org>
Commit-Queue: Colin Blundell <blundell@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1444102}