[ios] Clear WebState's UserData at the end of the destructor
Most of WebState's UserData objects have a pointer back to the
WebState. As base::SupportUserData is a super class of WebState
the base::SupportUserData::UserData were being destroyed after
the pointer to pimpl_ and saved_ have been destroyed and after
the vtable of WebState cleared. This made calling methods on
WebState really risky in WebState's UserData destructors.

To make this safer, call ClearAllUserData() at the end of the
destructor of WebStateImpl, ensuring that all the UserData are
destroyed before the object itself becomes invalid. This means
that the UserData object can keep a simple raw_ptr<WebState>
and be sure that it will be valid for their whole lifetime.

Bug: 409299519
Change-Id: Ib624d08cc1d2d446ec853fe1a733b4f238aceb79
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6439012
Reviewed-by: Olivier Robin <olivierrobin@chromium.org>
Commit-Queue: Sylvain Defresne <sdefresne@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1444128}